requirements.txt
numpy>=1.24
torch>=2.1
tqdm
fair-esm>=2.0.0
scikit-learn
